+++ forge-gui/res/cardsfolder/c/curse_of_leeches_leeching_lurker.txt
@@ -4,7 +4,7 @@ Types:Enchantment Aura Curse
 K:Enchant:Player
 SVar:AttachAILogic:Curse
 R:Event$ Transform | ValidCard$ Card.Self | ReplaceWith$ Attach | Description$ As this permanent transforms into CARDNAME, attach it to a player.
-SVar:Attach:DB$ Attach | Object$ Self | Chooser$ You | PlayerChoices$ Player | ChoiceTitle$ Choose a player to attach this Curse to | AILogic$ Curse
+SVar:Attach:DB$ Attach | Object$ Self | PlayerChoices$ Player | ChoiceTitle$ Choose a player to attach this Curse to | AILogic$ Curse
 T:Mode$ Phase | Phase$ Upkeep | ValidPlayer$ Player.EnchantedBy | TriggerZones$ Battlefield | Execute$ TrigDrain | TriggerDescription$ At the beginning of enchanted player's upkeep, they lose 1 life and you gain 1 life.
 SVar:TrigDrain:DB$ LoseLife | Defined$ TriggeredPlayer | LifeAmount$ 1 | SubAbility$ DBGainLife
 SVar:DBGainLife:DB$ GainLife | Defined$ You | LifeAmount$ 1
