+++ forge-gui/src/main/java/forge/gamemodes/net/client/NetGameController.java
@@ -227,26 +227,8 @@ public class NetGameController implements IGameController {
     }
 
     @Override
-    public boolean isUiSetToSkipPhase(final PlayerView turnPlayer, final PhaseType phase) {
-        // Never called on the client — host reads its own controller's cache
-        return false;
-    }
-
-    @Override
     public void setUiShouldSkipPhase(final PlayerView turnPlayer, final PhaseType phase, final boolean shouldSkip) {
         send(ProtocolMethod.setUiShouldSkipPhase, turnPlayer, phase, shouldSkip);
-    }
-
-    /** Sends only true entries; the host's empty-cache default already represents "don't skip". */
-    public void replayUiSkipPhases(final Iterable<PlayerView> allPlayers,
-                                   final java.util.function.BiPredicate<PlayerView, PhaseType> isSkipped) {
-        for (PlayerView p : allPlayers) {
-            for (PhaseType ph : PhaseType.values()) {
-                if (isSkipped.test(p, ph)) {
-                    send(ProtocolMethod.setUiShouldSkipPhase, p, ph, Boolean.TRUE);
-                }
-            }
-        }
     }
 
     private IMacroSystem macros;
